// projects/formints/formintB/frontend/src/types.ts
export interface Sale {
  id: number;
  total_amount: number;
  currency: string;
  date: string;
  time: string;
  order_type: string;
  status: string;
  table_number: number | null;
  delivery_type_id: number | null;
  delivery_zone_id?: number | null;
  delivery_address: string | null;
  employee_id: number | null;
  customer_id?: number | null;
  discount_code?: string | null;
  discount_amount?: number;
  payment_method?: string;
}